- updated to SumatraPDF v3.0

+++ sumatrapdf/src/mui/MuiButton.h
@@ -1,4 +1,4 @@
-/* Copyright 2013 the SumatraPDF project authors (see AUTHORS file).
+/* Copyright 2014 the SumatraPDF project authors (see AUTHORS file).
    License: Simplified BSD (see COPYING.BSD) */
 
 #ifndef Mui_h
@@ -39,7 +39,7 @@ public:
     void    SetStyles(Style *def, Style *mouseOver);
 
     WCHAR *         text;
-    size_t          textDx; // cached measured text width
+    int             textDx; // cached measured text width
 };
 
 // TODO: maybe should combine Button and ButtonVector into one?
